Send actual view height to webkit for window.innerHeight.
Currently only fixed height is sent over to webkit, which is to
avoid relayout when url bar is being scrolled.
And this fixed height is used also as the visible height and shown up
as window.innerheight, which is not correct.

This change sends the actual view height as the screen height to
webkit, so window.innerheight is correct.  This may potentially
help fixed elments in a page's bottom.  The relayout is still avoided
as before.

issue: 3396895
Change-Id: I0426a7d6dc76a823d8d26943a92689f49026fba8