WindowManagerService: Disable screen rotation animation under emulation.
Since there is no support for hardware GLES acceleration under
emulation, we disable the screen rotation animation when we detect
that we are in the emulator.

Note that ideally, we should only disable it when we detect that
GLES is implemented in software. However, I don't think there is
an easy way to test for this inside of the WindowManagerService.

Change-Id: Ic17719f7d0b14640bacb3b33b7b1a196006982ee