Fix the display priority of the airplane mode icon.
The airplane icon is used as a fallback when no
other radios are available *and* the AIRPLANE_MODE_ON system
setting is set to 1. (NetworkController now tracks changes to
the setting via ACTION_AIRPLANE_MODE_CHANGED broadcasts.)

This fixes wifi-only devices, for which the previous logic
did not correctly and consistently detect airplane mode.

Bug: 3489820
Change-Id: I226bb0e977f6a5fd3f2d829ba51ca54c5c910ef3