Contacts contract update for display photos.
This adds support for storing larger, hi-res photos in the contacts
system.  Thumbnails continue to be stored in the Photo.PHOTO column
as BLOBs, while hi-res versions of the photos are stored in the file
system and referenced via a photo file ID (with metadata stored in
a new table).

See go/large-photo-design for details.

Change-Id: Ie182eaa6a0330e2eca0a732f201f20a91e4dc584