Fix 5027468: prevent screen from turning back on while lockscreen in password mode.
This fixes a bug where LockScreen would turn back on due to the sequence of events
that ultimately cause PasswordUnlockScreen.onResume() to be called when the device powers off.

This is required because the lockscreen rebuilds itself before the screen turns off
so that it's ready to show immediately when the device comes back on.

Change-Id: I86a61cc3b1d77cf7f2a9a198051ea2bbe90607b2